+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for loading custom themes ​​from JSON
 - Migrated to CustomTkinter 6
 - Added optional support for using CTkTitleMenu instead of CTkMenuBar on Windows
+- Added a pre-installed Hello World Add-on that draws a simple picture on canvas
 
 ## 2.7.0 codename "Zviahel"
 - Fixed bugs